Rename ast and note
