+++ src/lib.rs
@@ -1,5 +1,5 @@
+mod ast;
 pub mod cli;
-mod md;
 mod note;
 
 use std::fs;
@@ -8,12 +8,12 @@ use anyhow::{anyhow, Context, Result};
 use markdown::mdast::Node;
 use markdown::{to_mdast, Constructs, ParseOptions};
 
+pub use crate::ast::builder;
+pub use crate::ast::pretty::pretty;
+pub use crate::ast::to_markdown::to_markdown;
 use crate::cli::config::Config;
-pub use crate::md::builder;
-pub use crate::md::pretty::pretty;
-pub use crate::md::serializer::to_markdown;
 use crate::note::metadata::Metadata;
-pub use crate::note::{from_note, to_note};
+pub use crate::note::{from_ast, to_ast};
 
 pub fn to_mdast_from_str(s: &str) -> Result<Node> {
     to_mdast(
@@ -30,9 +30,9 @@ pub fn to_mdast_from_str(s: &str) -> Result<Node> {
 }
 
 pub fn format(node: &Node) -> Result<String> {
-    let note = to_note(node)?;
+    let note = from_ast(node)?;
     let note = note.normalize();
-    let node = from_note(&note)?;
+    let node = to_ast(&note)?;
     to_markdown(&node)
 }
 
@@ -51,7 +51,7 @@ pub fn run(config: Config) -> Result<()> {
         }
 
         if config.note {
-            let note = to_note(&node)?;
+            let note = from_ast(&node)?;
             println!("{note:?}");
             return Ok(());
         }
