+++ src/note.rs
@@ -1,9 +1,9 @@
-mod from_note;
-mod to_note;
+mod from_ast;
+mod to_ast;
 
 pub mod flexible_date_time;
 pub mod metadata;
 pub mod model;
 
-pub use from_note::from_note;
-pub use to_note::to_note;
+pub use from_ast::from_ast;
+pub use to_ast::to_ast;
